# AquaSense/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AquaSense API",
    "version": "1.0.0",
    "description": "API для керування AquaSense системою"
  },
  "host": "localhost:3000",
  "schemes": ["http"],
...
  "paths": {
...
    "/sensors": {
      "post": {
        "tags": ["Sensors manager"],
        "summary": "Створити новий датчик",
        "description": "Додає новий датчик до бази даних",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Дані нового датчика",
            "schema": {
              "type": "object",
              "properties": {
                "aquarium_id": { "type": "integer" },
                "type": { "type": "string" },
                "value": { "type": "number", "format": "float" }
              },
              "required": ["aquarium_id", "type", "value"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Датчик успішно створено" },
          "500": { "description": "Помилка створення датчика" }
        }
      },
      "get": {
        "tags": ["Sensors manager"],
        "summary": "Отримати всі датчики",
        "description": "Повертає список усіх датчиків у системі",
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "aquarium_id": { "type": "integer" },
                  "type": { "type": "string" },
                  "value": { "type": "number", "format": "float" },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        }
      }
    },
    "/sensors/aquarium/{aquariumId}": {
      "get": {
        "tags": ["Sensors manager"],
        "summary": "Отримати всі датчики для акваріуму за його id",
        "description": "Повертає список датчиків, що належать до конкретного акваріума за його id",
        "parameters": [
          {
            "in": "path",
            "name": "aquariumId",
            "required": true,
            "type": "integer",
            "description": "id акваріума"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "aquarium_id": { "type": "integer" },
                  "type": { "type": "string" },
                  "value": { "type": "number", "format": "float" },
                  "created_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          },
          "404": {
            "description": "Акваріум не знайдено або в акваріумі немає датчиків"
          }
        }
      }
    },
    "/sensors/{id}": {
      "get": {
        "tags": ["Sensors manager"],
        "summary": "Отримати датчик за id",
        "description": "Повертає інформацію про датчик за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id датчика"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "aquarium_id": { "type": "integer" },
                "type": { "type": "string" },
                "value": { "type": "number", "format": "float" },
                "created_at": { "type": "string", "format": "date-time" }
              }
            }
          },
          "404": { "description": "Датчик не знайдено" }
        }
      },
      "put": {
        "tags": ["Sensors manager"],
        "summary": "Оновити інформацію про датчик",
        "description": "Оновлює інформацію про датчик за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id датчика"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Нові дані датчика",
            "schema": {
              "type": "object",
              "properties": {
                "type": { "type": "string" },
                "value": { "type": "number", "format": "float" }
              },
              "required": ["type", "value"]
            }
          }
        ],
        "responses": {
          "200": { "description": "Датчик успішно оновлено" },
          "404": { "description": "Датчик не знайдено" }
        }
      },
      "delete": {
        "tags": ["Sensors manager"],
        "summary": "Видалити датчик",
        "description": "Видаляє датчик за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id датчика"
          }
        ],
        "responses": {
          "200": { "description": "Датчик успішно видалено" },
          "404": { "description": "Датчик не знайдено" }
        }
      }
    },
    "/logs": {
      "post": {
        "tags": ["Logs manager"],
        "summary": "Створити новий лог",
        "description": "Додає новий лог до бази даних",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "Дані нового логу",
            "schema": {
              "type": "object",
              "properties": {
                "sensor_id": { "type": "integer" },
                "message": { "type": "string" }
              },
              "required": ["sensor_id", "message"]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Лог успішно створено"
          },
          "500": {
            "description": "Помилка створення логу"
          }
        }
      },
      "get": {
        "tags": ["Logs manager"],
        "summary": "Отримати всі логи",
        "description": "Повертає список усіх логів у системі",
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "sensor_id": { "type": "integer" },
                  "message": { "type": "string" },
                  "logged_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        }
      }
    },
    "/logs/sensor/{sensorId}": {
      "get": {
        "tags": ["Logs manager"],
        "summary": "Отримати всі логи для датчика за його id",
        "description": "Повертає список логів, пов'язаних із конкретним датчиком",
        "parameters": [
          {
            "in": "path",
            "name": "sensorId",
            "required": true,
            "type": "integer",
            "description": "id датчика"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "sensor_id": { "type": "integer" },
                  "message": { "type": "string" },
                  "logged_at": { "type": "string", "format": "date-time" }
                }
              }
            }
          },
          "404": {
            "description": "Логи для датчика не знайдено"
          }
        }
      },
      "delete": {
        "tags": ["Logs manager"],
        "summary": "Видалити всі логи для датчика за його id",
        "description": "Видаляє всі логи, пов'язані з конкретним датчиком",
        "parameters": [
          {
            "in": "path",
            "name": "sensorId",
            "required": true,
            "type": "integer",
            "description": "id датчика"
          }
        ],
        "responses": {
          "200": {
            "description": "Логи успішно видалено"
          },
          "404": {
            "description": "Логи для датчика не знайдено"
          }
        }
      }
    },
    "/logs/{id}": {
      "get": {
        "tags": ["Logs manager"],
        "summary": "Знайти лог за id",
        "description": "Повертає лог із вказаним id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id логу"
          }
        ],
        "responses": {
          "200": {
            "description": "Успішно",
            "schema": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "sensor_id": { "type": "integer" },
                "message": { "type": "string" },
                "logged_at": { "type": "string", "format": "date-time" }
              }
            }
          },
          "404": {
            "description": "Лог не знайдено"
          }
        }
      },
      "delete": {
        "tags": ["Logs manager"],
        "summary": "Видалити лог за його id",
        "description": "Видаляє лог за його id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "id логу"
          }
        ],
        "responses": {
          "200": {
            "description": "Лог успішно видалено"
          },
          "404": {
            "description": "Лог не знайдено"
          }
        }
      }
    }
  }
}